Rename files in preparation for a central monitor
This will use the same NodeMonitors as the edge monitor, but will be
driven from the ConfigDB instead of from k8s.
